fix: try to mitigate cand box not hide issue
It's related to pressing keys simultaneously (in short time)

https://www.autohotkey.com/docs/v2/misc/Threads.htm

When second key pressed, a new thread will be spawned, and the first
thread will be interrupted.

Assume the second key is `shift` and the cand box should hide - actually
the box hides indeed. However, then the first key's thread resumes, and
continue to run - then the box shows again.